Additional changes needed to use default RestOperationsSupplier
