+++ src/main/java/com/capitalone/dashboard/client/RestClientSettings.java
@@ -2,6 +2,11 @@ package com.capitalone.dashboard.client;
 
 public interface RestClientSettings {
 
-    int getRequestConnectTimeout();
-    int getRequestReadTimeout();
+    default int getRequestConnectTimeout() {
+        return 5000;
+    }
+
+    default int getRequestReadTimeout() {
+        return 60000;
+    }
 }
